requirements.txt
Django==5.1.7
djangorestframework==3.15.2
channels==4.2.0
daphne==4.1.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-ratelimit==4.1.0
psycopg2-binary==2.9.10
ollama==0.4.7
python-dotenv==1.0.1
